refactor: apply coding standards
